Dry Run option for server merge

+++ core/store/modules/cart/index.js
@@ -15,7 +15,7 @@ EventBus.$on('servercart-after-created', (event) => { // example stock check cal
   if (event.resultCode === 200) {
     console.log(`Server cart token after created = ${cartToken}`)
     rootStore.commit(types.SN_CART + '/' + types.CART_LOAD_CART_SERVER_TOKEN, cartToken)
-    rootStore.dispatch('cart/serverPull', { forceClientState: false }, { root: true })
+    rootStore.dispatch('cart/serverPull', { forceClientState: false, dryRun: !config.cart.server_merge_by_default }, { root: true })
   } else {
     let resultString = event.result ? _.toString(event.result) : null
     if (resultString && (resultString.indexOf(i18n.t('not authorized')) < 0 && resultString.indexOf('not authorized')) < 0) { // not respond to unathorized errors here
@@ -56,6 +56,7 @@ EventBus.$on('servercart-after-totals', (event) => { // example stock check call
 
 EventBus.$on('servercart-after-pulled', (event) => { // example stock check callback
   if (event.resultCode === 200) {
+    let diffLog = []
     let serverCartUpdateRequired = false
     let clientCartUpdateRequired = false
     let cartHasItems = false
@@ -69,26 +70,34 @@ EventBus.$on('servercart-after-pulled', (event) => { // example stock check call
 
       if (!serverItem) {
         console.log('No server item for ' + clientItem.sku)
-        rootStore.dispatch('cart/serverUpdateItem', {
-          sku: clientItem.parentSku && config.cart.setConfigurableProductOptions ? clientItem.parentSku : clientItem.sku,
-          qty: clientItem.qty,
-          product_option: clientItem.product_option
-        }, { root: true })
-        serverCartUpdateRequired = true
+        diffLog.push({ 'party': 'server', 'sku': clientItem.sku, 'status': 'no_item' })
+        if (!event.dry_run) {
+          rootStore.dispatch('cart/serverUpdateItem', {
+            sku: clientItem.parentSku && config.cart.setConfigurableProductOptions ? clientItem.parentSku : clientItem.sku,
+            qty: clientItem.qty,
+            product_option: clientItem.product_option
+          }, { root: true })
+          serverCartUpdateRequired = true
+        }
       } else if (serverItem.qty !== clientItem.qty) {
         console.log('Wrong qty for ' + clientItem.sku, clientItem.qty, serverItem.qty)
-        rootStore.dispatch('cart/serverUpdateItem', {
-          sku: clientItem.parentSku && config.cart.setConfigurableProductOptions ? clientItem.parentSku : clientItem.sku,
-          qty: clientItem.qty,
-          item_id: serverItem.item_id,
-          quoteId: serverItem.quote_id,
-          product_option: clientItem.product_option
-        }, { root: true })
-        serverCartUpdateRequired = true
+        diffLog.push({ 'party': 'server', 'sku': clientItem.sku, 'status': 'wrong_qty', 'client_qty': clientItem.qty, 'server_qty': serverItem.qty })
+        if (!event.dry_run) {
+          rootStore.dispatch('cart/serverUpdateItem', {
+            sku: clientItem.parentSku && config.cart.setConfigurableProductOptions ? clientItem.parentSku : clientItem.sku,
+            qty: clientItem.qty,
+            item_id: serverItem.item_id,
+            quoteId: serverItem.quote_id,
+            product_option: clientItem.product_option
+          }, { root: true })
+          serverCartUpdateRequired = true
+        }
       } else {
         console.log('Server and client items synced for ' + clientItem.sku) // here we need just update local item_id
         console.log('Updating server id to ', { sku: clientItem.sku, server_cart_id: serverItem.quote_id, server_item_id: serverItem.item_id, product_option: serverItem.product_option })
-        rootStore.dispatch('cart/updateItem', { product: { sku: clientItem.sku, server_cart_id: serverItem.quote_id, server_item_id: serverItem.item_id, product_option: serverItem.product_option } }, { root: true })
+        if (!event.dry_run) {
+          rootStore.dispatch('cart/updateItem', { product: { sku: clientItem.sku, server_cart_id: serverItem.quote_id, server_item_id: serverItem.item_id, product_option: serverItem.product_option } }, { root: true })
+        }
       }
     }
 
@@ -99,37 +108,44 @@ EventBus.$on('servercart-after-pulled', (event) => { // example stock check call
         })
         if (!clientItem) {
           console.log('No client item for ' + serverItem.sku)
+          diffLog.push({ 'party': 'client', 'sku': serverItem.sku, 'status': 'no_item' })
 
-          if (event.force_client_state) {
-            console.log('Removing item', serverItem.sku, serverItem.item_id)
-            serverCartUpdateRequired = true
-            rootStore.dispatch('cart/serverDeleteItem', {
-              sku: serverItem.sku,
-              item_id: serverItem.item_id,
-              quoteId: serverItem.quote_id
-            }, { root: true })
-          } else {
-            clientCartUpdateRequired = true
-            cartHasItems = true
-            rootStore.dispatch('product/single', { options: { sku: serverItem.sku }, setCurrentProduct: false, selectDefaultVariant: false }).then((product) => {
-              product.server_item_id = serverItem.item_id
-              product.qty = serverItem.qty
-              product.server_cart_id = serverItem.quote_id
-              if (serverItem.product_option) {
-                product.product_option = serverItem.product_option
-              }
-              rootStore.dispatch('cart/addItem', { productToAdd: product, forceServerSilence: true }).then(() => {
-              // rootStore.dispatch('cart/updateItem', { product: product })
+          if (!event.dry_run) {
+            if (event.force_client_state) {
+              console.log('Removing item', serverItem.sku, serverItem.item_id)
+              serverCartUpdateRequired = true
+              rootStore.dispatch('cart/serverDeleteItem', {
+                sku: serverItem.sku,
+                item_id: serverItem.item_id,
+                quoteId: serverItem.quote_id
+              }, { root: true })
+            } else {
+              clientCartUpdateRequired = true
+              cartHasItems = true
+              rootStore.dispatch('product/single', { options: { sku: serverItem.sku }, setCurrentProduct: false, selectDefaultVariant: false }).then((product) => {
+                product.server_item_id = serverItem.item_id
+                product.qty = serverItem.qty
+                product.server_cart_id = serverItem.quote_id
+                if (serverItem.product_option) {
+                  product.product_option = serverItem.product_option
+                }
+                rootStore.dispatch('cart/addItem', { productToAdd: product, forceServerSilence: true }).then(() => {
+                // rootStore.dispatch('cart/updateItem', { product: product })
+                })
               })
-            })
+            }
           }
         }
       }
     }
 
-    if ((!serverCartUpdateRequired || clientCartUpdateRequired) && cartHasItems) {
-      rootStore.dispatch('cart/refreshTotals')
+    if (!event.dry_run) {
+      if ((!serverCartUpdateRequired || clientCartUpdateRequired) && cartHasItems) {
+        rootStore.dispatch('cart/refreshTotals')
+      }
     }
+    EventBus.$emit('servercart-after-diff', { diffLog: diffLog, serverItems: serverItems, clientItems: clientItems, dryRun: event.dry_run, event: event }) // send the difflog
+    console.log('Server sync diff', diffLog)
   } else {
     console.error(event.result)
   }
